restructure component folders and redo layout to match design
