+++ src/components/AboutPage/AboutPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function AboutPage() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
