+++ src/components/Nav.jsx
@@ -1,8 +1,9 @@
 import React from 'react'
 
 export default function Nav(props) {
+
   return (
-    <header className='flex flex-row justify-between'>
+    <header className='flex flex-row justify-between mb-4'>
       <h1 className='text-xl'>Service Dogs Around Town</h1>
       <nav className='flex flex-row gap-4'>
         <button>Map</button>
